+++ src/net/sourceforge/kolmafia/request/PokemporiumRequest.java
@@ -53,7 +53,7 @@ public class PokemporiumRequest
 	private static final Pattern POKEDOLLAR_PATTERN = Pattern.compile( "([\\d,]+) 1,960 pok&eacute;dollar bills" );
 	public static final AdventureResult POKEDOLLAR = new AdventureResult( ItemPool.POKEDOLLAR_BILLS, 1, false ) {
 			@Override
-			public String getPluralName( int price )
+			public String getPluralName( long price )
 			{
 				return price == 1 ? "pok&eacute;dollar bill" : "pok&eacute;dollar bills";
 			}
